Use int64_t for pyvrp::Value instead of int32_t (#490)
* Switch to int64_t

* Switch to assert_allclose in older tests to avoid small precision issues

* Merge MAX_USER_VALUE and MAX_VALUE into a single constant

+++ pyvrp/cpp/Measure.h
@@ -15,7 +15,7 @@ namespace pyvrp
 #ifdef PYVRP_DOUBLE_PRECISION
 using Value = double;
 #else
-using Value = int32_t;
+using Value = int64_t;
 #endif
 
 enum class MeasureType
